fix: move sign pkg to gofakes3; use new lib/http
