+++ cmd/serve/s3/server.go
@@ -7,11 +7,11 @@ import (
 	"net/http"
 
 	"github.com/Mikubill/gofakes3"
-	"github.com/rclone/rclone/cmd/serve/httplib"
-	"github.com/rclone/rclone/cmd/serve/httplib/httpflags"
-	"github.com/rclone/rclone/cmd/serve/s3/signature"
+	"github.com/go-chi/chi/v5"
+	"github.com/go-chi/chi/v5/middleware"
 	"github.com/rclone/rclone/fs"
 	"github.com/rclone/rclone/fs/hash"
+	"github.com/rclone/rclone/lib/http/auth"
 	"github.com/rclone/rclone/vfs"
 	"github.com/rclone/rclone/vfs/vfsflags"
 )
@@ -22,13 +22,12 @@ type Options struct {
 	hostBucketMode bool
 	hashName       string
 	hashType       hash.Type
-	authPair       string
+	authPair       []string
 	noCleanup      bool
 }
 
 // Server is a s3.FileSystem interface
 type Server struct {
-	*httplib.Server
 	f       fs.Fs
 	vfs     *vfs.VFS
 	faker   *gofakes3.GoFakeS3
@@ -51,13 +50,22 @@ func newServer(ctx context.Context, f fs.Fs, opt *Options) *Server {
 		gofakes3.WithLogger(newLogger),
 		gofakes3.WithRequestID(rand.Uint64()),
 		gofakes3.WithoutVersioning(),
+		gofakes3.WithV4Auth(authlistResolver(opt.authPair)),
 	)
 
-	if opt.authPair != "" {
-		signature.LoadKeys(opt.authPair)
-	}
-
-	w.handler = w.authMiddleware(w.faker.Server())
-	w.Server = httplib.NewServer(w.handler, &httpflags.Opt)
+	w.handler = w.faker.Server()
+	// router.Route("/*", w.handler)
+	// w.Server = httplib.NewServer(w.handler, &httpflags.Opt)
 	return w
+}
+
+func (w *Server) Bind(router chi.Router) {
+	if m := auth.Auth(auth.Opt); m != nil {
+		router.Use(m)
+	}
+	router.Use(
+		middleware.SetHeader("Accept-Ranges", "bytes"),
+		middleware.SetHeader("Server", "rclone/"+fs.Version),
+	)
+	router.Handle("/*", w.handler)
 }
